Move the generators to a sub directory
