+++ gcovr/generator/sonarqube.py
@@ -7,7 +7,7 @@
 #
 # _____________________________________________________________________________
 #
-# Copyright (c) 2013-2021 the gcovr authors
+# Copyright (c) 2013-2020 the gcovr authors
 # Copyright (c) 2013 Sandia Corporation.
 # This software is distributed under the BSD License.
 # Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
@@ -18,7 +18,7 @@
 
 from lxml import etree
 
-from .utils import open_binary_for_writing, presentable_filename
+from ..utils import open_binary_for_writing, presentable_filename
 
 
 def print_sonarqube_report(covdata, output_file, options):
