+++ gcovr/generator/summary.py
@@ -7,7 +7,7 @@
 #
 # _____________________________________________________________________________
 #
-# Copyright (c) 2013-2021 the gcovr authors
+# Copyright (c) 2013-2020 the gcovr authors
 # Copyright (c) 2013 Sandia Corporation.
 # This software is distributed under the BSD License.
 # Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
@@ -18,7 +18,7 @@
 
 import sys
 
-from .utils import get_global_stats
+from ..utils import get_global_stats
 
 
 def print_summary(covdata):
